requirements.txt
asgiref==3.2.7
beautifulsoup4==4.8.2
bleach==3.1.5
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-appconf==1.0.4
django-bleach==0.6.1
django-cogwheels==0.3
django-compressor==2.4
django-crum==0.7.6
django-debug-toolbar==2.2
django-el-pagination==3.3.0
django-libsass==0.8
django-mathfilters==1.0.0
django-modelcluster==5.0.1
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
html5lib==1.0.1
idna==2.9
importlib-metadata==1.6.0
inflect==4.1.0
l18n==2018.5
libsass==0.20.0
packaging==20.3
Pillow==7.1.2
pyparsing==2.4.7
pytz==2020.1
rcssmin==1.0.6
requests==2.23.0
rjsmin==1.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.9
wagtail==2.9
wagtailmenus==3.0.1
webencodings==0.5.1
Willow==1.3
XlsxWriter==1.2.8
zipp==3.1.0
